pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.4",
    "gunicorn>=23.0.0",
    "langdetect>=1.0.9",
    "nlp-utils>=0.6.2",
    "nltk>=3.9.1",
    "numpy>=2.2.4",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.5",
]
